requirements.txt
-e .
bcrypt==3.1.7
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==3.3.2
distlib==0.3.9
enum34==1.1.10
filelock==3.2.1
hupper==1.10.3
idna==2.10
importlib-metadata==2.1.3
importlib-resources==3.3.1
ipaddress==1.0.23
Jinja2==2.11.3
MarkupSafe==2.1.2
netaddr==0.10.1
paramiko==2.12.0
parse==1.20.2
PasteDeploy==2.1.1
pathlib2==2.3.7.post1
plaster==1.0
plaster-pastedeploy==0.7
platformdirs==2.0.2
pycparser==2.21
PyNaCl==1.4.0
pyramid==2.0
python-hosts==0.4.7
python-magic==0.4.27
PyYAML==6.0 --only-binary :all:
repoze.lru==0.7
requests==2.27.1
singledispatch==3.7.0
six==1.17.0
tabulate==0.8.10
tinydb==3.15.2
translationstring==1.4
urllib3==1.26.20
venusian==2.1.0
virtualenv==20.15.1
WebOb==1.8.9
zipp==1.2.0
zope.deprecation==4.4.0
zope.interface==6.0
substanced==1.0.post1
gunicorn==20.1.0
git+https://github.com/Pylons/pyramid_chameleon
git+https://github.com/dholth/cryptacular
git+https://github.com/Pylons/walkabout
git+https://github.com/dlitz/python-pbkdf2
waitress
pyramid_tm
